// src/app/shared/models/approval.model.ts
export class Approval {
    userId: string;
    employeeId: string;
    leaveTypeCode: string;
    startDate: string;
    endDate: string;
    half: string;
    duration: string;
    updatedBy: string;
    updatedDate: string;
    remarks: string;
    unit: string;
    status: string;
    approverComment: string;
}