// src/app/shared/models/associate.model.ts
export interface Associate {
    employee_id: string;
    employee_name: string;
    designation: string;
    leave_type_code: string;
    leave_start_date: string;
    leave_end_date: string;
    half_type: string;
    duration: string;
    remark: string;
    status: string;
    status_description: string
}